IOS-3031 Fix bug for title button and make simple refactoring
